feat(api): parseIntegerParams function to ensure size and page are ints
