+++ packages/applications-service-api/src/routes/applications.js
@@ -5,7 +5,10 @@ const applicationsControllerV2 = require('../controllers/applications.v2');
 const { asyncRoute } = require('@pins/common/src/utils/async-route');
 const config = require('../lib/config');
 const { validateRequestWithOpenAPI } = require('../middleware/validator/openapi');
-const { normaliseArrayQueryParams } = require('../middleware/normaliseArrayQueryParams');
+const {
+	normaliseArrayQueryParams,
+	parseIntegerQueryParams
+} = require('../middleware/parseQueryParamProperties');
 
 const router = express.Router();
 
@@ -24,6 +27,7 @@ router.get('/download', applicationsController.getAllApplicationsDownload);
 router.get('/:caseReference', validateRequestWithOpenAPI, getApplicationsRoute);
 router.get(
 	'',
+	parseIntegerQueryParams(['page', 'size']),
 	normaliseArrayQueryParams(['stage', 'region', 'sector']),
 	validateRequestWithOpenAPI,
 	applicationsController.getAllApplications
